requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
distro==1.9.0
docstring_parser==0.18.0
faiss-cpu==1.13.2
filelock==3.28.0
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.3.0
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
